% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{crossValidate}
\alias{crossValidate}
\title{Perform cross validation}
\usage{
crossValidate(
  df,
  train.func,
  train.func.args = list(),
  colname.response = "response",
  k = 10,
  seed = 1,
  multi.core = F,
  verbose = 1
)
}
\arguments{
\item{df}{A dataframe containing the training features and response column.}

\item{train.func}{A training function.}

\item{train.func.args}{Arguments (as a list) that can be passed to the training function.}

\item{colname.response}{The column name of the response variable (i.e. training labels).}

\item{k}{Number of cross validation folds.}

\item{seed}{Random seed as in integer.}

\item{multi.core}{Use multiple cores?}

\item{verbose}{Show progress messages?}
}
\value{
A list containing for each CV fold, the output of the training function
}
\description{
Perform cross validation
}
