% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeatureImpHeatmap}
\alias{plotFeatureImpHeatmap}
\title{Make heatmap from a matrix of feature importances (for multiclass classification)}
\usage{
plotFeatureImpHeatmap(
  m = NULL,
  cv_out = NULL,
  top.n = 50,
  min.imp = NULL,
  invert.y = T,
  sort.features = F
)
}
\arguments{
\item{m}{A numeric matrix where columns represent the features and rows represent the labels}

\item{cv_out}{Alternative input to `m`. Cross validation output in the form of a list that has
'imp' object (e.g. cv_out[[1]]$imp)}

\item{top.n}{Top number of features to show}

\item{min.imp}{Features below this importance value are excluded from the plot}

\item{invert.y}{Invert the y-axis?
@param sort.features Order features alphabetically?}
}
\value{
A ggplot object
}
\description{
Make heatmap from a matrix of feature importances (for multiclass classification)
}
