% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTopFeatures}
\alias{plotTopFeatures}
\title{Make barplots from a matrix of feature importances (for multiclass classification)}
\usage{
plotTopFeatures(
  m = NULL,
  cv_out = NULL,
  top.n = 10,
  infer.feature.type = F,
  n.row = NULL,
  n.col = NULL
)
}
\arguments{
\item{m}{A numeric matrix where columns represent the features and rows represent the labels}

\item{cv_out}{Alternative input to `m`. Cross validation output in the form of a list that has
'imp' object (e.g. cv_out[[1]]$imp)}

\item{top.n}{Top number of features to show}

\item{infer.feature.type}{Determine the feature type based on the tag/prefix. Everything before
the first dot (.) is considered the feature tag}

\item{n.row}{Number of facet rows}

\item{n.col}{Number of facet columns}
}
\value{
A ggplot object
}
\description{
Make barplots from a matrix of feature importances (for multiclass classification)
}
