% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{univarFeatSel}
\alias{univarFeatSel}
\title{Performs pairwise testing and selects significant features}
\usage{
univarFeatSel(
  x,
  y,
  max.qvalue = 0.01,
  max.pvalue = NULL,
  sel.top.n.features = NULL,
  return.new.x = T,
  verbose = F
)
}
\arguments{
\item{x}{A matrix or dataframe of features}

\item{y}{A vector of class labels. For binary classification a logical vector. For
multiclass classification a factor.}

\item{max.qvalue}{qvalue threshold for keeping features}

\item{max.pvalue}{pvalue threshold for keeping features. Overrides max.qvalue}

\item{sel.top.n.features}{Limit the total number of features that are selected}

\item{return.new.x}{If TRUE, returns a feature matrix with the selected features. Else, a
character vector of the selected features}

\item{verbose}{Show progress messages?}
}
\value{
A vector of feature names if return.new.x=TRUE, else a feature matrix with the selected
features
}
\description{
For numerical variables, wilcoxon tests are performed. For categorical variables,
fisher exact tests are performed. The first factor level is assumed to be the negative outcome,
while the other levels are grouped together as the positive outcome. For example,
with the factor `as.factor(c('none','loh+pathogenic','deep_deletion'))`, 'none' is considered the
negative outcome.

When y is a factor (multiclass classification), multiple one-vs-rest pairwise tests (i.e. one for
each class label) are performed for each feature. A feature is kept if any of the pairwise tests
give a significant pvalue/qvalue.
}
